chore(layouts): don’t put layouts in separate directories
