+++ src/layouts/map.vue
@@ -50,5 +50,3 @@
     },
   };
 </script>
-
-<style src="./map.css"></style>
