--- src/layouts/map/map.css
@@ -1,4 +1,0 @@
-/* .app-main .v-main__wrap {
-  display: flex;
-  flex-direction: column;
-} */
